Fix reading data from netty byte buffers in the web socket provider
The problem with accessing the array-backed byte buffer was that it provided the entire array. Only a certain region of the array was necessary. A simple fix would be to copy the bytes directly, which would degrade performance. The more complex fix, which improves performance, would be to provide access to offset and length access to the encoder/decoder with the entire array.

The more complex fix has been applied. JsonObject and JsonArray now has support for handling offsets and lengths to access specific parts of a byte array during the encode/decode process.

This bug affected reading from text frames in web sockets.